pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "parallelproof"
version = "0.1.0"
description = "Multi-agent code optimizer with Tiger Agentic Postgres and Google Gemini"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.121.0",
    "uvicorn[standard]==0.38.0",
    "pydantic==2.12.4",
    "pydantic-settings==2.11.0",
    "asyncpg==0.30.0",
    "google-genai==1.49.0",
    "httpx==0.28.1",
    "websockets>=13.0,<15.0.dev0",
    "python-multipart==0.0.20",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.0",
    "pytest-asyncio==1.2.0",
    "black==25.9.0",
    "ruff==0.14.4",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["frontend*", "migrations*", "tests*", "scripts*"]
